Adds mainnet network Refactor logic
